+++ Packages/com.yontalane.core/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 1.0.8 - 2022.05.19
+
+**Added**
+
+**Changed**
+
+* Converted Collision Listener delegates to Unity Events.
+* Added Layer Mask filter to Collision Listener.
+
+**Fixed**
+
 ## 1.0.7 - 2022.05.19
 
 **Added**